feat : 처음 두장이 21일 경우 1.5배를 받는 기능 구현

+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,14 +1,14 @@
 package blackjack.domain;
 
 public class Dealer {
-    private static final int profit = 10000;
+    private static final double profit = 10000;
     private final PlayerCards myCards = new PlayerCards();
 
     public void addCard(Card card) {
         myCards.addCard(card);
     }
 
-    public int calculateResult() {
+    public double calculateResult() {
         return myCards.calculateResult();
     }
 
@@ -16,7 +16,7 @@ public class Dealer {
         return myCards.getCards();
     }
 
-    public int getProfit() {
+    public double getProfit() {
         return profit;
     }
